# PhotographyBack/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Mathis Petit",
        "description": "Documentation API pour le projet de photographie",
        "contact": {
            "email": "mathispetitfrpro@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/photos": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "Lister toutes les photos",
                "operationId": "8e878cf68237e684e4180bf0e1761116",
                "parameters": [
                    {
                        "name": "partner_id",
                        "in": "query",
                        "description": "Filtrer les photos par ID du partenaire",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des photos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Photo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Photos"
                ],
                "summary": "Créer une nouvelle photo",
                "operationId": "e21993f5b41f1c66ee580264d09c4cea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Titre de la photo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description de la photo"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Photo créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Photo"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/photos/{id}": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "Afficher une photo spécifique",
                "operationId": "6a1d12e7081fa8f2a91f99ec6a54f7b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la photo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la photo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Photo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Photo non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Photos"
                ],
                "summary": "Mettre à jour une photo",
                "operationId": "4710f9234345a7677c8bf4bd74955f19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la photo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Titre de la photo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description de la photo"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "category_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Photo"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Photos"
                ],
                "summary": "Supprimer une photo",
                "operationId": "c3a6f386d5b7a7af20cd331c2b411891",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la photo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Photo supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/photos/{id}/like": {
            "post": {
                "tags": [
                    "Photos"
                ],
                "summary": "Liker une photo",
                "operationId": "293c238fdac52c7ae89cb111e7f42530",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la photo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Photo likée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Photo likée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Photo déjà likée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/photos/{id}/unlike": {
            "post": {
                "tags": [
                    "Photos"
                ],
                "summary": "Retirer le like d'une photo",
                "operationId": "98af37c7cf4216f1021f3d97b13fc427",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la photo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like retiré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Like retiré avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Like non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category",
                "description": "Category model",
                "properties": {
                    "id": {
                        "description": "ID of the category",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the category",
                        "type": "string",
                        "example": "Landscape"
                    },
                    "created_at": {
                        "description": "Timestamp when the category was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-04T12:34:56.000000Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the category was updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-05T12:34:56.000000Z"
                    },
                    "photos": {
                        "description": "List of photos belonging to the category",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            },
            "Partner": {
                "title": "Partner",
                "description": "Modèle représentant un partenaire photographe dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique du partenaire",
                        "type": "integer"
                    },
                    "bio": {
                        "description": "Biographie du partenaire",
                        "type": "string"
                    },
                    "website": {
                        "description": "Site web du partenaire",
                        "type": "string"
                    },
                    "social_links": {
                        "description": "Liens vers les réseaux sociaux du partenaire",
                        "type": "string"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur associé au partenaire",
                        "type": "integer"
                    },
                    "photos": {
                        "description": "Liste des photos associées à ce partenaire",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "Photo": {
                "title": "Photo",
                "description": "Modèle représentant une photo dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique de la photo",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Titre de la photo",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description de la photo",
                        "type": "string"
                    },
                    "image_path": {
                        "description": "Chemin vers l'image de la photo",
                        "type": "string"
                    },
                    "likes": {
                        "description": "Nombre de likes que la photo a reçu",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur ayant téléchargé la photo",
                        "type": "integer"
                    },
                    "categories": {
                        "description": "Liste des catégories associées à cette photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "likedByUsers": {
                        "description": "Liste des utilisateurs ayant liké la photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Modèle représentant un utilisateur dans l'application",
                "properties": {
                    "id": {
                        "description": "ID unique de l'utilisateur",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom de l'utilisateur",
                        "type": "string"
                    },
                    "email": {
                        "description": "Adresse email de l'utilisateur",
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "description": "Rôle de l'utilisateur (user, partner, admin)",
                        "type": "string"
                    },
                    "profile_photo_url": {
                        "description": "URL de la photo de profil de l'utilisateur",
                        "type": "string"
                    },
                    "photos": {
                        "description": "Liste des photos associées à cet utilisateur",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    },
                    "partner": {
                        "$ref": "#/components/schemas/Partner"
                    },
                    "likedPhotos": {
                        "description": "Liste des photos likées par l'utilisateur",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}